refactor: Split lexer.rs into focused submodules
Split the monolithic 608-line lexer.rs file into smaller, focused
modules for better maintainability and clarity:

- lexer/mod.rs (177 lines): Main Lexer struct and tokenization API
- lexer/keywords.rs (193 lines): SQL keyword recognition (184 keywords)
- lexer/numbers.rs (135 lines): Number parsing with scientific notation
- lexer/strings.rs (36 lines): String literal parsing with escapes
- lexer/identifiers.rs (61 lines): Regular and delimited identifiers
- lexer/operators.rs (54 lines): Multi-character operator recognition

All 649 parser tests pass without modification. Token stream interface
remains unchanged, ensuring backward compatibility.
